Added class names and styles to elements
created a div called food-container and ac lass called food list for
the ul found in index.html. Used to center and style the list with
styles.css. A description class was also created and added to pizza.html.
The class will be used for the other html files in recipes directory
to style those descriptions as well. General styling was also added for
images, h1, and body that all html files will have in common.